+++ cps-ri/src/main/java/org/onap/cps/spi/entities/FragmentExtract.java
@@ -24,7 +24,7 @@ public interface FragmentExtract {
 
     Long getId();
 
-    Integer getAnchorId();
+    Long getAnchorId();
 
     String getXpath();
